refactor MCP folder
