+++ openhands/mcp/__init__.py
@@ -1,8 +1,12 @@
-from openhands.mcp.mcp import (
+from openhands.mcp.client import MCPClient
+from openhands.mcp.tool import (
     BaseTool,
-    MCPClient,
+    MCPClientTool,
+)
+from openhands.mcp.utils import (
     convert_mcp_clients_to_tools,
     create_mcp_clients,
+    fetch_mcp_tools_from_config,
 )
 
 __all__ = [
@@ -10,4 +14,6 @@ __all__ = [
     'convert_mcp_clients_to_tools',
     'create_mcp_clients',
     'BaseTool',
+    'MCPClientTool',
+    'fetch_mcp_tools_from_config',
 ]
