+++ tests/unit/test_mcp_timeout.py
@@ -4,7 +4,7 @@ from unittest import mock
 import pytest
 
 from openhands.core.config.mcp_config import MCPConfig, MCPSSEConfig, MCPStdioConfig
-from openhands.mcp.mcp import MCPClient, create_mcp_clients, fetch_mcp_tools_from_config
+from openhands.mcp import MCPClient, create_mcp_clients, fetch_mcp_tools_from_config
 
 
 @pytest.mark.asyncio
